fix: FDMiData failed on Set()
- 初歩的なミス

+++ Packages/tech.gyoku.fdmi.core/Runtime/Data/FDMiFloat.cs
@@ -20,8 +20,8 @@ namespace FDMi.core
             TriggerCallbacks();
         }
         
-        public override void Set(short i) => Set(i);
-        public override void Set(sbyte i) => Set(i);
+        public override void Set(short i) => Set((float)i);
+        public override void Set(sbyte i) => Set((float)i);
         public override float GetFloat() => data[0];
         public override short GetShort() => (short)(data[0]);
         public override sbyte GetSByte() => (sbyte)(data[0]);
